fix handling dynamic path
because getCurrentScriptSource executes at runtime in the browser, it can simply look at __webpack_public_path__ which means that it should always be up-to-date, even if an app dynamically changes it. however, the app has to execute first before the plugin, so that the function is looking at a public path that may have been changed by the app.