chore: update type-fest version
npm 7 throws an error when installing packages if the versions of peerDependencies are not compatible.

Since type-fest is a utility package, you will likely use many versions, so you should include as many versions as possible to avoid breaking DX.

Now, this package uses only Except, SetRequired, and LiteralUnion in type-fest, so it is expected to work normally in 0.17.0 or later.

### Reference
https://github.blog/2021-02-02-npm-7-is-now-generally-available/#peer-dependencies
https://github.com/sindresorhus/type-fest/releases/tag/v0.6.0
https://github.com/sindresorhus/type-fest/releases/tag/v0.8.0
https://github.com/sindresorhus/type-fest/releases/tag/v0.17.0