Update package.json
There is a new version of webpack-dev-server with a semver 5.0.0

This change would make it possible to update webpack-dev-server to the latest version 5.0.0 without needing to force the update.

It would also allow users to test with the latest version of webpack-dev-server in development. I installed in 2 projects bundled with webpack 5 and React 18, and I have have not issues with hot reload.